+++ R/eta_squared.R
@@ -153,7 +153,9 @@ eta_squared.aov <- function(model, partial = TRUE, ci = 0.9, ...) {
     }
   }
 
-  class(out) <- c(ifelse(isTRUE(partial), "partial_eta_squared", "eta_squared"), class(out))
+  class(out) <- c(ifelse(isTRUE(partial), "partial_eta_squared", "eta_squared"),
+                  "effectsize_table",
+                  class(out))
   out
 }
 
@@ -190,7 +192,7 @@ eta_squared.anova <- function(model, partial = TRUE, ci = 0.9, ...) {
     )
   )
 
-  class(out) <- c("partial_eta_squared", class(out))
+  class(out) <- c("partial_eta_squared", "effectsize_table", class(out))
   out
 }
 
@@ -222,7 +224,7 @@ eta_squared.aovlist <- function(model, partial = TRUE, ci = 0.9, ...) {
               ci = ci)
   )
 
-  class(out) <- c("partial_eta_squared", class(out))
+  class(out) <- c("partial_eta_squared", "effectsize_table", class(out))
   out
 }