+++ R/omega_squared.R
@@ -74,7 +74,9 @@ omega_squared.aov <- function(model, partial = TRUE, ci = 0.9, ...) {
     }
   }
 
-  class(out) <- c(ifelse(isTRUE(partial), "partial_omega_squared", "omega_squared"), class(out))
+  class(out) <- c(ifelse(isTRUE(partial), "partial_omega_squared", "omega_squared"),
+                  "effectsize_table",
+                  class(out))
   out
 }
 
@@ -110,7 +112,7 @@ omega_squared.anova <- function(model, partial = TRUE, ci = 0.9, ...) {
     )
   )
 
-  class(out) <- c("partial_omega_squared", class(out))
+  class(out) <- c("partial_omega_squared", "effectsize_table", class(out))
   out
 }
 
@@ -142,7 +144,7 @@ omega_squared.aovlist <- function(model, partial = TRUE, ci = 0.9, ...) {
                   ci = ci)
   )
 
-  class(out) <- c("partial_omega_squared", class(out))
+  class(out) <- c("partial_omega_squared", "effectsize_table", class(out))
   out
 }